<firstname>/Add script fonctions

+++ script-docker.sh
@@ -1,35 +1,38 @@
 #!/bin/bash
 
-# Mettre à jour la liste des paquets
-sudo apt-get update
+if dpkg-query -l docker-compose >/dev/null 2>&1; then
 
-# Installer Docker Compose
-sudo apt install docker-compose
+    echo " * Docker-compose"
+    echo "     est déjà installé avec la version $(docker-compose -v)."
+else
 
-# Supprimer les anciennes versions de Docker
-sudo apt install docker.io docker-compose
+    # Installer Docker Compose
+    sudo apt install docker-compose
 
-# Mettre à jour à nouveau la liste des paquets
-sudo apt update
+    # Supprimer les anciennes versions de Docker
+    sudo apt install docker.io docker-compose
 
-# Installer les dépendances nécessaires
-sudo apt update
+    # Mettre à jour à nouveau la liste des paquets
+    sudo apt update
 
-sudo apt install apt-transport-https ca-certificates curl software-properties-common
+    # Installer les dépendances nécessaires
+    sudo apt update
 
+    sudo apt install apt-transport-https ca-certificates curl software-properties-common
 
-# Ajouter la clé GPG du site de Docker
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
+    # Ajouter la clé GPG du site de Docker
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
 
-# Configurer le référentiel stable de Docker :
-echo "deb [arch=amd64 signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+    # Configurer le référentiel stable de Docker :
+    echo "deb [arch=amd64 signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list >/dev/null
 
-# Installer Docker Engine :
-sudo apt update
+    # Installer Docker Engine :
+    sudo apt update
 
-sudo apt install docker-ce docker-ce-cli containerd.io
+    sudo apt install docker-ce docker-ce-cli containerd.io
 
-# Installer Docker Compose
-sudo apt install docker-compose
+    # Installer Docker Compose
+    sudo apt install docker-compose
 
-sudo apt update
+    sudo apt update
+fi
